requirements.txt
httpx>=0.27.0
trafilatura>=1.7.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
langdetect>=1.0.9
pymupdf>=1.24.9
rapidfuzz>=3.9.6
simhash>=2.0.1
python-dotenv>=1.0.1

qdrant-client>=1.9.1
sentence-transformers>=3.0.1
transformers>=4.43.3
torch>=2.3.1

fastapi>=0.112.0
uvicorn[standard]>=0.30.5
selenium>=4.15.0
